requirements.txt
google-trans-new
streamlit
gtts
mtranslate
pyautogui
openpyxl
python-docx